docker/judge/requirements.txt
flask
flask-mysql
flask-cors
PyYAML
